Refactor to generic store syntax
