+++ stores/index.js
@@ -1,0 +1,6 @@
+exports.billa = require("./billa");
+exports.dm = require("./dm");
+exports.hofer = require("./hofer");
+exports.lidl = require("./lidl");
+exports.mpreis = require("./mpreis");
+exports.spar = require("./spar");
